requirements.txt
ccxt==4.2.84
python-telegram-bot==21.4
python-dotenv==1.0.1